Renaming
Also moved some common functions in monty-31 to a utils crate
