+++ Cargo.toml
@@ -22,7 +22,7 @@ members = [
     "maybe-rayon",
     "mersenne-31",
     "monolith",
-    "monty-31-bit-field",
+    "monty-31",
     "poseidon",
     "poseidon2",
     "rescue",
