+++ monty-31/src/lib.rs
@@ -5,6 +5,8 @@
 mod data_traits;
 mod extension;
 mod monty_31_bit_field;
+mod utils;
 
 pub use data_traits::*;
 pub use monty_31_bit_field::*;
+pub use utils::*;
